# mbms/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounting/financial-periods": {
      "get": {
        "operationId": "FinancialPeriodsController_list",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "FinancialPeriodsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFinancialPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/accounting/financial-periods/{id}/close": {
      "post": {
        "operationId": "FinancialPeriodsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseFinancialPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/accounting/journal-entries": {
      "get": {
        "operationId": "JournalEntriesController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[financialPeriodId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[status]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "JournalEntriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJournalEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/accounting/journal-entries/{id}/post": {
      "post": {
        "operationId": "JournalEntriesController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/accounting/reconciliations": {
      "get": {
        "operationId": "ReconciliationController_list",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "ReconciliationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/accounting/reconciliations/{id}/reconcile": {
      "post": {
        "operationId": "ReconciliationController_reconcile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/ap/invoices": {
      "get": {
        "operationId": "AccountsPayableController_listInvoices",
        "parameters": [
          {
            "name": "filter[companyId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[supplierId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[status]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "AccountsPayableController_createInvoice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ap/invoices/{id}/submit-for-approval": {
      "post": {
        "operationId": "AccountsPayableController_submitForApproval",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/invoices/{id}/approve": {
      "post": {
        "operationId": "AccountsPayableController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/payments": {
      "post": {
        "operationId": "AccountsPayableController_recordPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordSupplierPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/credit-notes": {
      "post": {
        "operationId": "AccountsPayableController_createCreditNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierCreditNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/aging": {
      "get": {
        "operationId": "AccountsPayableController_aging",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/statement": {
      "get": {
        "operationId": "AccountsPayableController_statement",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "supplierId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/ap/payment-schedule": {
      "get": {
        "operationId": "AccountsPayableController_paymentSchedule",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Morise MBMS API",
    "description": "Morise Holdings Limited Business Management System — Phase 1 API. Base path /api/v1. Staff endpoints take a `staff` bearer token from POST /identity/auth/login; Customer Storefront endpoints take a `customer` bearer token from POST /customer-portal/auth/login.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "staff": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "customer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFinancialPeriodDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "periodName": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "periodName",
          "startDate",
          "endDate"
        ]
      },
      "CloseFinancialPeriodDto": {
        "type": "object",
        "properties": {
          "retainedEarningsAccountId": {
            "type": "string"
          }
        }
      },
      "JournalEntryItemDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "debitAmount": {
            "type": "number",
            "minimum": 0
          },
          "creditAmount": {
            "type": "number",
            "minimum": 0
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "accountId"
        ]
      },
      "CreateJournalEntryDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "financialPeriodId": {
            "type": "string"
          },
          "entryDate": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "entryType": {
            "type": "string",
            "enum": [
              "standard",
              "adjusting"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntryItemDto"
            }
          }
        },
        "required": [
          "companyId",
          "financialPeriodId",
          "entryDate",
          "items"
        ]
      },
...
      "CreateReconciliationDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "asOfDate": {
            "type": "string"
          },
          "statementBalance": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "accountId",
          "asOfDate",
          "statementBalance"
        ]
      },
...
      "InvoiceItemDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "description",
          "unitPrice"
        ]
      },
      "CreateSupplierInvoiceDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "invoiceDate": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "taxAmount": {
            "type": "number",
            "minimum": 0
          },
          "apAccountId": {
            "type": "string"
          },
          "expenseAccountId": {
            "type": "string"
          },
          "financialPeriodId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemDto"
            }
          }
        },
        "required": [
          "companyId",
          "supplierId",
          "invoiceDate",
          "dueDate",
          "currency",
          "apAccountId",
          "expenseAccountId",
          "financialPeriodId",
          "items"
        ]
      },
      "RecordSupplierPaymentDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "currency": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "bankAccountId": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "supplierId",
          "paymentDate",
          "amount",
          "currency",
          "bankAccountId",
          "invoiceId"
        ]
      },
      "CreateSupplierCreditNoteDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "creditDate": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "reason": {
            "type": "string"
          },
          "appliedToInvoiceId": {
            "type": "string"
          },
          "offsetAccountId": {
            "type": "string"
          },
          "financialPeriodId": {
            "type": "string"
          }
        },
        "required": [
          "companyId",
          "supplierId",
          "creditDate",
          "amount",
          "offsetAccountId",
          "financialPeriodId"
        ]
      },
...
    }
  }
}